// src/interface/client/clientInterface.ts
import { ILoginUserData } from "../user/userInterface";

export interface IClientData {
  id?: number;
  email?: string;
  weekendDays?: string;
  contractN?: string | null;
  contractTagline?: string | null;
  code?: string;
  name?: string;
  timezone?: string;
  currency?: string;
  country?: string;
  isActive?: boolean;
  startDate?: Date;
  endDate?: Date;
  autoUpdateEndDate?: number;
  timeSheetStartDay?: number;
  approvalEmail?: string;
  isShowPrices?: boolean;
  isShowCostCenter?: boolean;
  isShowCatalogueNo?: boolean;
  titreDeConge?: string;
  isResetBalance?: boolean;
  startMonthBack?: number;
  medicalEmailSubmission?: string;
  medicalEmailToday?: string;
  medicalEmailMonthly?: string;
  isShowNSS?: boolean;
  isShowCarteChifa?: boolean;
  isShowSalaryInfo?: boolean;
  isShowRotation?: boolean;
  isShowBalance?: boolean;
  isCountCR?: boolean;
  loginUserData?: ILoginUserData;
  logo?: string | null;
  stampLogo?: string | null;
  segment?: string;
  subSegment?: string;
  bonusType?: string;
  address?: string;
  createdAt?: Date | string;
  createdBy?: number;
  updatedAt?: Date | string;
  updatedBy?: number;
  deletedAt?: Date | null | string;
}

export interface IClientResponseData {
  id?: number;
  code?: string;
  country?: string;
  isActive?: boolean;
  startDate?: Date;
  endDate?: Date;
  autoUpdateEndDate?: number;
  timeSheetStartDay?: number;
  approvalEmail?: string;
  isShowPrices?: boolean;
  isShowCostCenter?: boolean;
  isShowCatalogueNo?: boolean;
  titreDeConge?: string;
  isResetBalance?: boolean;
  startMonthBack?: number;
  medicalEmailSubmission?: string;
  medicalEmailToday?: string;
  medicalEmailMonthly?: string;
  isShowNSS?: boolean;
  isShowCarteChifa?: boolean;
  isShowSalaryInfo?: boolean;
  isShowRotation?: boolean;
  isShowBalance?: boolean;
  isCountCR?: boolean;
  logo?: string;
  segment?: string;
  subSegment?: string;
  bonusType?: string;
  createdAt?: Date | string;
  createdBy?: number;
  updatedAt?: Date | string;
  updatedBy?: number;
  deletedAt?: Date | null | string;
  loginUserData?: ILoginUserData;
}